BLogList Expansion --> 4.21 step 9: delete only by owner

+++ part4/blogList/backend/controllers/blogs.js
@@ -37,6 +37,18 @@ blogsRouter.post('/', async (req, res) => {
 
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  console.log(decodedToken)
+  if (!decodedToken.id)
+    return res.status(401).json({ error: 'invalid token' })
+
+  const blogToDelete = await Blog.findById(req.params.id)
+  if (!blogToDelete)
+    return res.status(404).json({ error: 'blog doesnt exist' })
+
+  if (blogToDelete.user.toString() !== decodedToken.id.toString())
+    return res.status(403).json({ error: 'forbidden: not your blog' })
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
